Converting from smallint to int so validation of DeniedValuesAttribute work. Improving tests so that they only spawn a single database container

+++ src/FairPlayCombinedSln/FairPlayCombined.DataAccess/Models/FairPlayDating/Activity.cs
@@ -18,7 +18,7 @@ namespace FairPlayCombined.DataAccess.Models.FairPlayDatingSchema;
 public partial class Activity
 {
     [Key]
-    public short ActivityId { get; set; }
+    public int ActivityId { get; set; }
 
     [Required]
     [StringLength(50)]
