+++ src/FairPlayCombinedSln/FairPlayCombined.DataAccess/Models/FairPlayDating/UserActivity.cs
@@ -24,9 +24,9 @@ public partial class UserActivity
     [Required]
     public string ApplicationUserId { get; set; }
 
-    public short ActivityId { get; set; }
+    public int ActivityId { get; set; }
 
-    public short FrequencyId { get; set; }
+    public int FrequencyId { get; set; }
 
     [ForeignKey("ActivityId")]
     [InverseProperty("UserActivity")]
