+++ src/FairPlayCombinedSln/FairPlayCombined.Models/FairPlayDating/EyesColor/EyesColorModel.cs
@@ -10,7 +10,7 @@ namespace FairPlayCombined.Models.FairPlayDating.EyesColor
 {
     public class EyesColorModel : IListModel
     {
-        public short? EyesColorId { get; set; }
+        public int? EyesColorId { get; set; }
         public string? Name { get; set; }
     }
 }
