+++ src/FairPlayCombinedSln/FairPlayCombinedDb/FairPlayDating/Frequency.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [FairPlayDating].[Frequency]
 (
-	[FrequencyId] SMALLINT NOT NULL CONSTRAINT PK_Frequency PRIMARY KEY IDENTITY, 
+	[FrequencyId] INT NOT NULL CONSTRAINT PK_Frequency PRIMARY KEY IDENTITY, 
     [Name] NVARCHAR(50) NOT NULL
 )
